fixed: request 관련 오류 해결

+++ hobbyhobby-community/hobbyhobby-storage/src/main/java/org/v1/UserCommunityEntityRepository.java
@@ -9,12 +9,12 @@ import java.util.List;
 public class UserCommunityEntityRepository implements UserCommunityRepository {
 
     @Override
-    public void appendUserToCommunity(Long userId, Long communityId) {
+    public void appendUserToCommunity(Long userId, Integer communityId) {
 
     }
 
     @Override
-    public void removeUserFromCommunity(Long userId, Long communityId) {
+    public void removeUserFromCommunity(Long userId, Integer communityId) {
 
     }
 }
